chore: Fix YAKS tests for Camel 4.5.0
- Update Apache Camel runtime to 4.5.0
- Use Citrus 4.2.0
- Update YAKS 0.19.0
