+++ pom.xml
@@ -65,9 +65,9 @@
         <camel.version>4.5.0</camel.version>
         <camel.k.crds.version>2.2.0</camel.k.crds.version>
 
-        <citrus.version>4.1.0</citrus.version>
+        <citrus.version>4.2.0</citrus.version>
         <cucumber.version>7.16.1</cucumber.version>
-        <yaks.version>0.17.1</yaks.version>
+        <yaks.version>0.19.0</yaks.version>
 
         <!-- Versions used inside Kamelets (add them also to the dependencyManagement section and the groovy script below) -->
         <!-- These properties must keep this same format "version.<groupId>.<artifactId>" -->
