+++ tests/camel-kamelets-itest/src/test/resources/avro/User.avsc
