+++ tests/camel-kamelets-itest/src/test/resources/avro/avro-binary-source-pipe.yaml
