+++ tests/camel-kamelets-itest/src/test/resources/avro/avro-deserialize-pipe.yaml
