+++ tests/camel-kamelets-itest/src/test/resources/avro/avro-serialize-pipe.yaml
