+++ tests/camel-kamelets-itest/src/test/resources/avro/avro-x-struct-sink-pipe.yaml
