+++ tests/camel-kamelets-itest/src/test/resources/aws/ddb/verifyItems.groovy
@@ -15,10 +15,18 @@
  * limitations under the License.
  */
 
-System.properties['citrus.mail.marshaller.type'] = "JSON"
+package aws.ddb
 
-mail()
-  .server('mail-server')
-  .port(22222)
-  .knownUsers(['${email}:${username}:${password}'])
-  .autoStart(true)
+$(repeatOnError()
+    .until('i > ${maxRetryAttempts}')
+    .actions(new org.citrusframework.TestAction() {
+        @Override
+        void execute(org.citrusframework.context.TestContext context) {
+            try {
+                assert context.getVariable('aws.ddb.items').equals(amazonDDBClient.scan(b -> b.tableName(context.getVariable('aws.ddb.tableName')))?.items()?.toListString())
+            } catch (AssertionError e) {
+                throw new org.citrusframework.exceptions.CitrusRuntimeException("AWS DDB item verification failed", e)
+            }
+        }
+    })
+)
