+++ tests/camel-kamelets-itest/src/test/resources/aws/s3/amazonS3Client.groovy
@@ -1,3 +1,4 @@
+package aws.s3
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements. See the NOTICE file distributed with
@@ -18,36 +19,20 @@
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider
 import software.amazon.awssdk.regions.Region
-import software.amazon.awssdk.services.dynamodb.DynamoDbClient
-import software.amazon.awssdk.services.dynamodb.model.AttributeDefinition
-import software.amazon.awssdk.services.dynamodb.model.KeySchemaElement
-import software.amazon.awssdk.services.dynamodb.model.KeyType
-import software.amazon.awssdk.services.dynamodb.model.ProvisionedThroughput
-import software.amazon.awssdk.services.dynamodb.model.ScalarAttributeType
+import software.amazon.awssdk.services.s3.S3Client
 
-DynamoDbClient amazonDDBClient = DynamoDbClient
+S3Client s3 = S3Client
         .builder()
-        .endpointOverride(URI.create("${YAKS_TESTCONTAINERS_LOCALSTACK_DYNAMODB_LOCAL_URL}"))
+        .endpointOverride(URI.create("${YAKS_TESTCONTAINERS_LOCALSTACK_S3_URL}"))
         .credentialsProvider(StaticCredentialsProvider.create(
                 AwsBasicCredentials.create(
                         "${YAKS_TESTCONTAINERS_LOCALSTACK_ACCESS_KEY}",
                         "${YAKS_TESTCONTAINERS_LOCALSTACK_SECRET_KEY}")
         ))
+        .forcePathStyle(true)
         .region(Region.of("${YAKS_TESTCONTAINERS_LOCALSTACK_REGION}"))
         .build()
 
-amazonDDBClient.createTable(b -> {
-        b.tableName("${aws.ddb.tableName}")
-        b.keySchema(
-                KeySchemaElement.builder().attributeName("id").keyType(KeyType.HASH).build(),
-        )
-        b.attributeDefinitions(
-                AttributeDefinition.builder().attributeName("id").attributeType(ScalarAttributeType.N).build(),
-        )
-        b.provisionedThroughput(
-                ProvisionedThroughput.builder()
-                        .readCapacityUnits(1L)
-                        .writeCapacityUnits(1L).build())
-})
+s3.createBucket(b -> b.bucket("${aws.s3.bucketNameOrArn}"))
 
-return amazonDDBClient
+return s3
