+++ tests/camel-kamelets-itest/src/test/resources/aws/s3/aws-s3-credentials.properties
