+++ tests/camel-kamelets-itest/src/test/resources/aws/s3/aws-s3-source-secret-conf.feature
@@ -16,15 +16,15 @@ Feature: AWS S3 Source - secret based config
       | aws-s3-credentials.properties | citrus:encodeBase64(citrus:readFile(aws-s3-credentials.properties)) |
 
   Scenario: Verify AWS-S3 Kamelet to log binding
+    # Create AWS-S3 client
+    Given New Camel context
+    Given load to Camel registry amazonS3Client.groovy
     # Create binding
     Given create labels on Kubernetes secret aws-s3-source-credentials
       | camel.apache.org/kamelet               | aws-s3-source |
       | camel.apache.org/kamelet.configuration | aws-s3-credentials |
     Given load Camel K integration aws-s3-to-log-secret-based.groovy
     Then Camel K integration aws-s3-to-log-secret-based should be running
-    # Create AWS-S3 client
-    Given New Camel context
-    Given load to Camel registry amazonS3Client.groovy
     # Verify Kamelet source
     Given Camel exchange message header CamelAwsS3Key="${aws.s3.key}"
     Given send Camel exchange to("aws2-s3://${aws.s3.bucketNameOrArn}?amazonS3Client=#amazonS3Client") with body: ${aws.s3.message}
