+++ tests/camel-kamelets-itest/src/test/resources/aws/s3/aws-s3-source-uri-conf.feature
@@ -12,12 +12,12 @@ Feature: AWS S3 Source - URI based config
     Given start LocalStack container
 
   Scenario: Verify AWS-S3 Kamelet to log binding
-    # Create binding
-    Given load Camel K integration aws-s3-to-log-uri-based.groovy
-    Then Camel K integration aws-s3-to-log-uri-based should be running
     # Create AWS-S3 client
     Given New Camel context
     Given load to Camel registry amazonS3Client.groovy
+    # Create binding
+    Given load Camel K integration aws-s3-to-log-uri-based.groovy
+    Then Camel K integration aws-s3-to-log-uri-based should be running
     # Verify Kamelet source
     Given Camel exchange message header CamelAwsS3Key="${aws.s3.key}"
     Given send Camel exchange to("aws2-s3://${aws.s3.bucketNameOrArn}?amazonS3Client=#amazonS3Client") with body: ${aws.s3.message}
