+++ tests/camel-kamelets-itest/src/test/resources/aws/s3/aws-s3-to-knative-broker.yaml
