+++ tests/camel-kamelets-itest/src/test/resources/aws/s3/knative-to-log.yaml
