+++ tests/camel-kamelets-itest/src/test/resources/aws/s3/yaks-config.yaml
@@ -39,15 +39,15 @@ config:
       - name: CITRUS_TYPE_CONVERTER
         value: camel
     resources:
-      - ../../amazonS3Client.groovy
-      - ../../aws-s3-credentials.properties
-      - ../../aws-s3-to-log-uri-based.groovy
-      - ../../aws-s3-to-log-secret-based.groovy
-      - ../../aws-s3-uri-pipe.yaml
-      - ../../aws-s3-to-http.yaml
-      - ../../aws-s3-to-knative-channel.yaml
-      - ../../aws-s3-to-knative-broker.yaml
-      - ../../knative-channel-to-log.yaml
+      - amazonS3Client.groovy
+      - aws-s3-credentials.properties
+      - aws-s3-to-log-uri-based.groovy
+      - aws-s3-to-log-secret-based.groovy
+      - aws-s3-uri-pipe.yaml
+      - aws-s3-to-http.yaml
+      - aws-s3-to-knative-channel.yaml
+      - aws-s3-to-knative-broker.yaml
+      - knative-channel-to-log.yaml
     cucumber:
       tags:
         - "not @ignored and not @knative"
