+++ tests/camel-kamelets-itest/src/test/resources/citrus-application.properties
@@ -8,8 +8,8 @@ citrus.type.converter=camel
 yaks.cluster.type=local
 yaks.camelk.max.attempts=10
 
-yaks.jbang.camel.version=4.4.0-SNAPSHOT
-yaks.jbang.kamelets.version=4.4.0-SNAPSHOT
+yaks.jbang.camel.version=4.5.0
+yaks.jbang.kamelets.version=4.6.0-SNAPSHOT
 
 yaks.jbang.kamelets.local.dir=../../../kamelets
 yaks.jbang.camel.dump.integration.output=true
