+++ tests/camel-kamelets-itest/src/test/resources/earthquake/earthquake-to-http.yaml
