+++ tests/camel-kamelets-itest/src/test/resources/kafka/kafka-sink-pipe.yaml
