+++ tests/camel-kamelets-itest/src/test/resources/kafka/kafka-source-pipe.yaml
