+++ tests/camel-kamelets-itest/src/test/resources/kafka/yaks-config.yaml
@@ -37,8 +37,8 @@ config:
         - name: INTEGRATION_LOGS
           level: INFO
     resources:
-      - ../kafka-source-pipe.yaml
-      - ../kafka-sink-pipe.yaml
+      - kafka-source-pipe.yaml
+      - kafka-sink-pipe.yaml
   dump:
     enabled: true
     failedOnly: true
