+++ tests/camel-kamelets-itest/src/test/resources/mail/timer-to-mail.yaml
