+++ tests/camel-kamelets-itest/src/test/resources/mail/yaks-config.yaml
@@ -37,8 +37,8 @@ config:
           artifactId: citrus-mail
           version: "@citrus.version@"
     resources:
-      - ../mail-server.groovy
-      - ../timer-to-mail.yaml
+      - mail-server.groovy
+      - timer-to-mail.yaml
   dump:
     enabled: true
     failedOnly: true
