+++ tests/camel-kamelets-itest/src/test/resources/openapi/openapi.json
