+++ tests/camel-kamelets-itest/src/test/resources/openapi/pet.json
