+++ tests/camel-kamelets-itest/src/test/resources/openapi/rest-openapi-sink-pipe.yaml
