+++ tests/camel-kamelets-itest/src/test/resources/openapi/yaks-config.yaml
@@ -33,9 +33,9 @@ config:
         - name: INTEGRATION_LOGS
           level: INFO
     resources:
-      - ../pet.json
-      - ../openapi.json
-      - ../rest-openapi-sink-pipe.yaml
+      - pet.json
+      - openapi.json
+      - rest-openapi-sink-pipe.yaml
   dump:
     enabled: true
     failedOnly: true
