+++ tests/camel-kamelets-itest/src/test/resources/salesforce/direct-to-salesforce-update-pipe.yaml
