+++ tests/camel-kamelets-itest/src/test/resources/salesforce/salesforce-credentials.properties
