+++ tests/camel-kamelets-itest/src/test/resources/salesforce/timer-to-salesforce-pipe.yaml
