+++ tests/camel-kamelets-itest/src/test/resources/timer/timer-to-http-pipe.yaml
