+++ tests/camel-kamelets-itest/src/test/resources/timer/yaks-config.yaml
@@ -33,9 +33,9 @@ config:
         - name: INTEGRATION_LOGS
           level: INFO
     resources:
-      - ../timer-to-http-pipe.yaml
-      - ../timer-to-http.groovy
-      - ../timer-to-http.properties
+      - timer-to-http-pipe.yaml
+      - timer-to-http.groovy
+      - timer-to-http.properties
   dump:
     enabled: true
     failedOnly: true
