+++ tests/camel-kamelets-itest/src/test/resources/transformation/data-type-action-pipe.yaml
