+++ tests/camel-kamelets-itest/src/test/resources/transformation/extract-field-action-pipe.yaml
