+++ tests/camel-kamelets-itest/src/test/resources/transformation/insert-field-action-pipe.yaml
