+++ tests/camel-kamelets-itest/src/test/resources/transformation/yaks-config.yaml
@@ -33,9 +33,9 @@ config:
         - name: INTEGRATION_LOGS
           level: INFO
     resources:
-      - ../insert-field-action-pipe.yaml
-      - ../extract-field-action-pipe.yaml
-      - ../data-type-action-pipe.yaml
+      - insert-field-action-pipe.yaml
+      - extract-field-action-pipe.yaml
+      - data-type-action-pipe.yaml
   dump:
     enabled: true
     failedOnly: true
